Fix a 32-bit compilation failure

+++ src/or/routerparse.h
@@ -101,7 +101,7 @@ typedef struct {
   time_t when;
 } dumped_desc_t;
 
-EXTERN(size_t, len_descs_dumped);
+EXTERN(uint64_t, len_descs_dumped);
 EXTERN(smartlist_t *, descs_dumped);
 STATIC int routerstatus_parse_guardfraction(const char *guardfraction_str,
                                             networkstatus_t *vote,
